refactor: pull long event card out into own component
